requirements.txt
# Core dependencies
face_recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# Configuration and environment
python-dotenv==1.0.0

# Progress and CLI utilities
tqdm==4.66.1
click==8.1.7

# Development and testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
mypy==1.6.1

# Optional: GPU acceleration for face detection
# Uncomment if you have CUDA installed
# dlib-gpu==19.24.2

# Optional: Additional video format support
# Uncomment if needed
# av==10.0.0
# imageio[ffmpeg]==2.31.5

# Documentation (optional)
# sphinx==7.2.6
# sphinx-rtd-theme==1.3.0